Update README and enhance resume management with user authorization checks

+++ resumes/resume_routers.py
@@ -70,7 +70,7 @@ def create_resume(resume_form: ResumeRequest, session: Session = Depends(get_ses
     if not current_user:
         raise UnauthorizedException(detail='You must be logged in to create a resume')
 
-    resume = rs.create_resume(resume_form, session)
+    resume = rs.create_resume(resume_form, session, current_user)
 
     if not resume:
         raise NotFoundException(detail='Resume could not be created')
@@ -83,6 +83,12 @@ def update_resume(id: int, resume_form: ResumeUpdate, session: Session = Depends
 
     if not current_user:
         raise UnauthorizedException(detail='You must be logged in to update a resume')
+    
+    if not rs.get_resume_by_id(session, id):
+        raise NotFoundException(detail='Resume not found')
+    
+    if current_user.id.user_id != rs.get_resume_by_id(session, id).user_id:
+        raise UnauthorizedException(detail='You are not authorized to update this resume')
 
     resume = rs.update_resume(id, resume_form, session)
 
@@ -98,6 +104,12 @@ def delete_resume(id: int, session: Session = Depends(get_session), current_user
 
     if not current_user:
         raise UnauthorizedException(detail='You must be logged in to delete a resume')
+    
+    if not rs.get_resume_by_id(session, id):
+        raise NotFoundException(detail='Resume not found')
+    
+    if current_user.id.user_id != rs.get_resume_by_id(session, id).user_id:
+        raise UnauthorizedException(detail='You are not authorized to delete this resume')
     
     resume = rs.delete_resume(id, session)
